Remove pubPath and dartVersion from DAP config
See #2783.
