+++ src/debug/web_debug_impl.ts
@@ -42,6 +42,6 @@ export class WebDebugSession extends FlutterDebugSession {
 		}
 
 		// TODO: Attach?
-		return new WebRun(isAttach ? RunMode.Attach : RunMode.Run, args.pubPath, args.cwd, appArgs, { envOverrides: args.env, toolEnv: this.toolEnv }, args.webDaemonLogFile, logger, (url) => this.exposeUrl(url), this.maxLogLineLength);
+		return new WebRun(isAttach ? RunMode.Attach : RunMode.Run, args.dartSdkPath, args.cwd, appArgs, { envOverrides: args.env, toolEnv: this.toolEnv }, args.webDaemonLogFile, logger, (url) => this.exposeUrl(url), this.maxLogLineLength);
 	}
 }
